Divide components into PictureCard and PicturesComponents
